Hold the reading apart from what it came to
The conclusion carried the reading it came from, so `Open` could be
built beside a reading that says the opposite of it — the sentence "the
model divides this position no way" was held by a factory getting a
condition right, which is the discipline this protocol was rebuilt to
stop needing. Now a reading is completed and an answer derived from it,
and the pair is what a caller gets:

    LocalInspection(reading, Divided | Open | Blocked)

`Open` has nothing of its own to contradict, and the pair refuses to be
made from a reading short of the rules, so no value carrying the
contradiction exists.

`Divided` carries the completeness through to the axis. Not making a
consumer for it yet and dropping the producer's answer are different
things: a class off a set arrived at from part of the rules is a value
the model singled out, and a rule that went unread may yet refuse it, so
the denominator is the one the model states rather than one every class
of which is known to be inhabited. Dropped at the axis, the only thing
left to read would be the classes, which say nothing about what was read
to get them.

`Rules.of` loses its `default`, so a declaration kind added later has to
be classified rather than arriving as a value quietly said to have no
rules written about it.

And what stopped a reading is told by what it recognised rather than by
how many positions the rule mentions. `validPair(left, right)` names two
and may say nothing about how they stand against each other; counting
them projected to a word that tells the author their rule compares one
position with another. A comparison of one position with another is its
own answer; anything else is a form this reading does not take apart.
79 positions in souther-examples move from the first to the second, and
`unsupported_partition_shape` drops to the 7 that are comparisons.

Three more, found by review of the same code:

The classes left by the crossing are in the order the type declares
them. Ordered by the rule instead, `== Won || == Prospecting` and `==
Prospecting || == Won` were one partition written two ways.

A position that can hold none of its classes keeps them. Nothing left is
not an empty partition — it is a value nothing can build, refused where
the declaration is (E1013) — and read as one it came back as a position
the model divides no way.

A rule on the declaration is said ahead of what a body's comparison came
to. Both are true and one line is written: the rule on the declaration
is the one whose being read would give the position an axis, and a
comparison relating it to another position divides it no way however
well it is read.
