+++ souther-compiler/src/main/java/souther/compiler/partition/Axis.java
@@ -1,6 +1,7 @@
 package souther.compiler.partition;
 
 import souther.compiler.types.Type;
+import souther.compiler.values.AdmissibleSet;
 
 import java.util.Collection;
 import java.util.List;
@@ -31,6 +32,11 @@ import java.util.Set;
  *                 path is how a reason came to be recovered by string match. A reason travels with
  *                 the position or it is a reason about whatever the strings happened to pair it
  *                 with.
+ * @param read     how much of what the rules say about this position's values was read. Carried
+ *                 because it qualifies the classes and nothing else says it: a class off a set
+ *                 arrived at from part of the rules is a value the model singled out, and a rule
+ *                 that went unread may yet refuse it — so {@link #eligible} is the denominator the
+ *                 model states and not one every class of which is known to be inhabited
  * @param unread   a rule about this position's own values that the local reading did not take in,
  *                 or null where it read them all. Carried for the same reason {@link #pending} is,
  *                 and kept apart from it because the two are lifted by different work and one
@@ -38,23 +44,27 @@ import java.util.Set;
  *                 a rule about what is inside describes that same stop from the other end
  */
 public record Axis(AxisId id, NumericTerm term, Type type, List<PartitionClass> classes,
-                   List<Cut> cuts, Set<String> excluded, StructuralInspection.Pending pending,
-                   BlockReason unread) {
+                   List<Cut> cuts, Set<String> excluded, AdmissibleSet.Completeness read,
+                   StructuralInspection.Pending pending, BlockReason unread) {
 
     public Axis {
         classes = List.copyOf(classes);
         cuts = List.copyOf(cuts);
         excluded = Set.copyOf(excluded);
+        if (read == null) {
+            throw new IllegalArgumentException(
+                    "a position with no account of what was read about its values");
+        }
     }
 
     public Axis(AxisId id, NumericTerm term, Type type, List<PartitionClass> classes,
                 List<Cut> cuts, Set<String> excluded) {
-        this(id, term, type, classes, cuts, excluded, null, null);
+        this(id, term, type, classes, cuts, excluded, AdmissibleSet.READ_IN_FULL, null, null);
     }
 
     public Axis(AxisId id, NumericTerm term, Type type, List<PartitionClass> classes,
                 List<Cut> cuts) {
-        this(id, term, type, classes, cuts, Set.of(), null, null);
+        this(id, term, type, classes, cuts, Set.of(), AdmissibleSet.READ_IN_FULL, null, null);
     }
 
     /**
@@ -65,8 +75,9 @@ public record Axis(AxisId id, NumericTerm term, Type type, List<PartitionClass> 
      * reading ran to the end and found nothing, and what stopped one where it did not.
      */
     public static Axis pendingAt(AxisId id, NumericTerm term, Type type,
+                                 AdmissibleSet.Completeness read,
                                  StructuralInspection.Pending found, BlockReason unread) {
-        return new Axis(id, term, type, List.of(), List.of(), Set.of(), found, unread);
+        return new Axis(id, term, type, List.of(), List.of(), Set.of(), read, found, unread);
     }
 
     /**
@@ -80,12 +91,12 @@ public record Axis(AxisId id, NumericTerm term, Type type, List<PartitionClass> 
      * as one the model divides no way.
      */
     public Axis measuredAt(AxisId id, NumericTerm term) {
-        return new Axis(id, term, type, classes, cuts, excluded, pending, unread);
+        return new Axis(id, term, type, classes, cuts, excluded, read, pending, unread);
     }
 
     /** The same position, with what a body's rules divided it into and the lines they drew. */
     public Axis carrying(List<PartitionClass> classes, List<Cut> cuts) {
-        return new Axis(id, term, type, classes, cuts, excluded, pending, unread);
+        return new Axis(id, term, type, classes, cuts, excluded, read, pending, unread);
     }
 
     /** Where the value this axis is about sits, which is where a row is walked to before the term is
@@ -104,7 +115,7 @@ public record Axis(AxisId id, NumericTerm term, Type type, List<PartitionClass> 
         Set<String> here = ids.stream().filter(id -> classOf(id) != null)
                 .collect(java.util.stream.Collectors.toCollection(java.util.LinkedHashSet::new));
         return here.isEmpty() ? this
-                : new Axis(id, term, type, classes, cuts, here, pending, unread);
+                : new Axis(id, term, type, classes, cuts, here, read, pending, unread);
     }
 
     /**
